ml-service/requirements.txt
torch>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

